+++ sdk/elasticsans/arm-elasticsan/src/elasticSanManagement.ts
@@ -11,9 +11,15 @@ import * as coreRestPipeline from "@azure/core-rest-pipeline";
 import {
   PipelineRequest,
   PipelineResponse,
-  SendRequest
+  SendRequest,
 } from "@azure/core-rest-pipeline";
 import * as coreAuth from "@azure/core-auth";
+import {
+  SimplePollerLike,
+  OperationState,
+  createHttpPoller,
+} from "@azure/core-lro";
+import { createLroSpec } from "./lroImpl";
 import {
   OperationsImpl,
   SkusImpl,
@@ -22,7 +28,7 @@ import {
   VolumesImpl,
   PrivateEndpointConnectionsImpl,
   PrivateLinkResourcesImpl,
-  VolumeSnapshotsImpl
+  VolumeSnapshotsImpl,
 } from "./operations";
 import {
   Operations,
@@ -32,9 +38,15 @@ import {
   Volumes,
   PrivateEndpointConnections,
   PrivateLinkResources,
-  VolumeSnapshots
+  VolumeSnapshots,
 } from "./operationsInterfaces";
-import { ElasticSanManagementOptionalParams } from "./models";
+import * as Parameters from "./models/parameters";
+import * as Mappers from "./models/mappers";
+import {
+  ElasticSanManagementOptionalParams,
+  RestoreVolumeOptionalParams,
+  RestoreVolumeResponse,
+} from "./models";
 
 export class ElasticSanManagement extends coreClient.ServiceClient {
   $host: string;
@@ -50,7 +62,7 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
   constructor(
     credentials: coreAuth.TokenCredential,
     subscriptionId: string,
-    options?: ElasticSanManagementOptionalParams
+    options?: ElasticSanManagementOptionalParams,
   ) {
     if (credentials === undefined) {
       throw new Error("'credentials' cannot be null");
@@ -65,10 +77,10 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
     }
     const defaults: ElasticSanManagementOptionalParams = {
       requestContentType: "application/json; charset=utf-8",
-      credential: credentials
+      credential: credentials,
     };
 
-    const packageDetails = `azsdk-js-arm-elasticsan/1.0.1`;
+    const packageDetails = `azsdk-js-arm-elasticsan/1.1.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`
@@ -78,20 +90,21 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
       ...defaults,
       ...options,
       userAgentOptions: {
-        userAgentPrefix
+        userAgentPrefix,
       },
       endpoint:
-        options.endpoint ?? options.baseUri ?? "https://management.azure.com"
+        options.endpoint ?? options.baseUri ?? "https://management.azure.com",
     };
     super(optionsWithDefaults);
 
     let bearerTokenAuthenticationPolicyFound: boolean = false;
     if (options?.pipeline && options.pipeline.getOrderedPolicies().length > 0) {
-      const pipelinePolicies: coreRestPipeline.PipelinePolicy[] = options.pipeline.getOrderedPolicies();
+      const pipelinePolicies: coreRestPipeline.PipelinePolicy[] =
+        options.pipeline.getOrderedPolicies();
       bearerTokenAuthenticationPolicyFound = pipelinePolicies.some(
         (pipelinePolicy) =>
           pipelinePolicy.name ===
-          coreRestPipeline.bearerTokenAuthenticationPolicyName
+          coreRestPipeline.bearerTokenAuthenticationPolicyName,
       );
     }
     if (
@@ -101,7 +114,7 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
       !bearerTokenAuthenticationPolicyFound
     ) {
       this.pipeline.removePolicy({
-        name: coreRestPipeline.bearerTokenAuthenticationPolicyName
+        name: coreRestPipeline.bearerTokenAuthenticationPolicyName,
       });
       this.pipeline.addPolicy(
         coreRestPipeline.bearerTokenAuthenticationPolicy({
@@ -111,9 +124,9 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
             `${optionsWithDefaults.endpoint}/.default`,
           challengeCallbacks: {
             authorizeRequestOnChallenge:
-              coreClient.authorizeRequestOnClaimChallenge
-          }
-        })
+              coreClient.authorizeRequestOnClaimChallenge,
+          },
+        }),
       );
     }
     // Parameter assignments
@@ -121,7 +134,7 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
 
     // Assigning values to Constant parameters
     this.$host = options.$host || "https://management.azure.com";
-    this.apiVersion = options.apiVersion || "2023-01-01";
+    this.apiVersion = options.apiVersion || "2024-06-01-preview";
     this.operations = new OperationsImpl(this);
     this.skus = new SkusImpl(this);
     this.elasticSans = new ElasticSansImpl(this);
@@ -142,7 +155,7 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
       name: "CustomApiVersionPolicy",
       async sendRequest(
         request: PipelineRequest,
-        next: SendRequest
+        next: SendRequest,
       ): Promise<PipelineResponse> {
         const param = request.url.split("?");
         if (param.length > 1) {
@@ -156,9 +169,115 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
           request.url = param[0] + "?" + newParams.join("&");
         }
         return next(request);
-      }
+      },
     };
     this.pipeline.addPolicy(apiVersionPolicy);
+  }
+
+  /**
+   * Restore Soft Deleted Volumes.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param elasticSanName The name of the ElasticSan.
+   * @param volumeGroupName The name of the VolumeGroup.
+   * @param volumeName The name of the Volume.
+   * @param options The options parameters.
+   */
+  async beginRestoreVolume(
+    resourceGroupName: string,
+    elasticSanName: string,
+    volumeGroupName: string,
+    volumeName: string,
+    options?: RestoreVolumeOptionalParams,
+  ): Promise<
+    SimplePollerLike<
+      OperationState<RestoreVolumeResponse>,
+      RestoreVolumeResponse
+    >
+  > {
+    const directSendOperation = async (
+      args: coreClient.OperationArguments,
+      spec: coreClient.OperationSpec,
+    ): Promise<RestoreVolumeResponse> => {
+      return this.sendOperationRequest(args, spec);
+    };
+    const sendOperationFn = async (
+      args: coreClient.OperationArguments,
+      spec: coreClient.OperationSpec,
+    ) => {
+      let currentRawResponse: coreClient.FullOperationResponse | undefined =
+        undefined;
+      const providedCallback = args.options?.onResponse;
+      const callback: coreClient.RawResponseCallback = (
+        rawResponse: coreClient.FullOperationResponse,
+        flatResponse: unknown,
+      ) => {
+        currentRawResponse = rawResponse;
+        providedCallback?.(rawResponse, flatResponse);
+      };
+      const updatedArgs = {
+        ...args,
+        options: {
+          ...args.options,
+          onResponse: callback,
+        },
+      };
+      const flatResponse = await directSendOperation(updatedArgs, spec);
+      return {
+        flatResponse,
+        rawResponse: {
+          statusCode: currentRawResponse!.status,
+          body: currentRawResponse!.parsedBody,
+          headers: currentRawResponse!.headers.toJSON(),
+        },
+      };
+    };
+
+    const lro = createLroSpec({
+      sendOperationFn,
+      args: {
+        resourceGroupName,
+        elasticSanName,
+        volumeGroupName,
+        volumeName,
+        options,
+      },
+      spec: restoreVolumeOperationSpec,
+    });
+    const poller = await createHttpPoller<
+      RestoreVolumeResponse,
+      OperationState<RestoreVolumeResponse>
+    >(lro, {
+      restoreFrom: options?.resumeFrom,
+      intervalInMs: options?.updateIntervalInMs,
+      resourceLocationConfig: "location",
+    });
+    await poller.poll();
+    return poller;
+  }
+
+  /**
+   * Restore Soft Deleted Volumes.
+   * @param resourceGroupName The name of the resource group. The name is case insensitive.
+   * @param elasticSanName The name of the ElasticSan.
+   * @param volumeGroupName The name of the VolumeGroup.
+   * @param volumeName The name of the Volume.
+   * @param options The options parameters.
+   */
+  async beginRestoreVolumeAndWait(
+    resourceGroupName: string,
+    elasticSanName: string,
+    volumeGroupName: string,
+    volumeName: string,
+    options?: RestoreVolumeOptionalParams,
+  ): Promise<RestoreVolumeResponse> {
+    const poller = await this.beginRestoreVolume(
+      resourceGroupName,
+      elasticSanName,
+      volumeGroupName,
+      volumeName,
+      options,
+    );
+    return poller.pollUntilDone();
   }
 
   operations: Operations;
@@ -170,3 +289,38 @@ export class ElasticSanManagement extends coreClient.ServiceClient {
   privateLinkResources: PrivateLinkResources;
   volumeSnapshots: VolumeSnapshots;
 }
+// Operation Specifications
+const serializer = coreClient.createSerializer(Mappers, /* isXml */ false);
+
+const restoreVolumeOperationSpec: coreClient.OperationSpec = {
+  path: "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ElasticSan/elasticSans/{elasticSanName}/volumegroups/{volumeGroupName}/volumes/{volumeName}/restore",
+  httpMethod: "POST",
+  responses: {
+    200: {
+      bodyMapper: Mappers.Volume,
+    },
+    201: {
+      bodyMapper: Mappers.Volume,
+    },
+    202: {
+      bodyMapper: Mappers.Volume,
+    },
+    204: {
+      bodyMapper: Mappers.Volume,
+    },
+    default: {
+      bodyMapper: Mappers.ErrorResponse,
+    },
+  },
+  queryParameters: [Parameters.apiVersion],
+  urlParameters: [
+    Parameters.$host,
+    Parameters.subscriptionId,
+    Parameters.resourceGroupName,
+    Parameters.elasticSanName,
+    Parameters.volumeGroupName,
+    Parameters.volumeName,
+  ],
+  headerParameters: [Parameters.accept],
+  serializer,
+};