+++ sdk/elasticsans/arm-elasticsan/src/models/index.ts
@@ -414,6 +414,8 @@ export interface VolumeGroupProperties {
    * NOTE: This property will not be serialized. It can only be populated by the server.
    */
   readonly privateEndpointConnections?: PrivateEndpointConnection[];
+  /** The retention policy for the deleted volumes. */
+  deleteRetentionPolicy?: DeleteRetentionPolicy;
 }
 
 /** The encryption settings on the volume group. */
@@ -469,6 +471,14 @@ export interface VirtualNetworkRule {
   action?: Action;
 }
 
+/** Response for Delete Retention Policy object */
+export interface DeleteRetentionPolicy {
+  /** The flag to enable/disable retention policy. */
+  enabled?: boolean;
+  /** The number of days to retain the volume after deletion. */
+  days?: number;
+}
+
 /** Volume Group request. */
 export interface VolumeGroupUpdate {
   /** The identity of the resource. */
@@ -487,6 +497,8 @@ export interface VolumeGroupUpdateProperties {
   encryptionProperties?: EncryptionProperties;
   /** A collection of rules governing the accessibility from specific network locations. */
   networkAcls?: NetworkRuleSet;
+  /** The retention policy for the deleted volumes. */
+  deleteRetentionPolicy?: DeleteRetentionPolicy;
 }
 
 /** Volume response properties. */
@@ -734,6 +746,11 @@ export interface VolumesDeleteHeaders {
   location?: string;
 }
 
+/** Defines headers for ElasticSanManagement_restoreVolume operation. */
+export interface ElasticSanManagementRestoreVolumeHeaders {
+  location?: string;
+}
+
 /** Defines headers for PrivateEndpointConnections_delete operation. */
 export interface PrivateEndpointConnectionsDeleteHeaders {
   location?: string;
@@ -751,7 +768,7 @@ export enum KnownOrigin {
   /** System */
   System = "system",
   /** UserSystem */
-  UserSystem = "user,system"
+  UserSystem = "user,system",
 }
 
 /**
@@ -768,7 +785,7 @@ export type Origin = string;
 /** Known values of {@link ActionType} that the service accepts. */
 export enum KnownActionType {
   /** Internal */
-  Internal = "Internal"
+  Internal = "Internal",
 }
 
 /**
@@ -785,7 +802,7 @@ export enum KnownSkuName {
   /** Premium locally redundant storage */
   PremiumLRS = "Premium_LRS",
   /** Premium zone redundant storage */
-  PremiumZRS = "Premium_ZRS"
+  PremiumZRS = "Premium_ZRS",
 }
 
 /**
@@ -801,7 +818,7 @@ export type SkuName = string;
 /** Known values of {@link SkuTier} that the service accepts. */
 export enum KnownSkuTier {
   /** Premium Tier */
-  Premium = "Premium"
+  Premium = "Premium",
 }
 
 /**
@@ -830,7 +847,13 @@ export enum KnownProvisioningStates {
   /** Updating */
   Updating = "Updating",
   /** Deleting */
-  Deleting = "Deleting"
+  Deleting = "Deleting",
+  /** Deleted */
+  Deleted = "Deleted",
+  /** Restoring */
+  Restoring = "Restoring",
+  /** SoftDeleting */
+  SoftDeleting = "SoftDeleting",
 }
 
 /**
@@ -845,7 +868,10 @@ export enum KnownProvisioningStates {
  * **Pending** \
  * **Creating** \
  * **Updating** \
- * **Deleting**
+ * **Deleting** \
+ * **Deleted** \
+ * **Restoring** \
+ * **SoftDeleting**
  */
 export type ProvisioningStates = string;
 
@@ -858,7 +884,7 @@ export enum KnownPrivateEndpointServiceConnectionStatus {
   /** Failed */
   Failed = "Failed",
   /** Rejected */
-  Rejected = "Rejected"
+  Rejected = "Rejected",
 }
 
 /**
@@ -882,7 +908,7 @@ export enum KnownCreatedByType {
   /** ManagedIdentity */
   ManagedIdentity = "ManagedIdentity",
   /** Key */
-  Key = "Key"
+  Key = "Key",
 }
 
 /**
@@ -902,7 +928,7 @@ export enum KnownPublicNetworkAccess {
   /** Enabled */
   Enabled = "Enabled",
   /** Disabled */
-  Disabled = "Disabled"
+  Disabled = "Disabled",
 }
 
 /**
@@ -915,6 +941,24 @@ export enum KnownPublicNetworkAccess {
  */
 export type PublicNetworkAccess = string;
 
+/** Known values of {@link XMsAccessSoftDeletedResources} that the service accepts. */
+export enum KnownXMsAccessSoftDeletedResources {
+  /** True */
+  True = "true",
+  /** False */
+  False = "false",
+}
+
+/**
+ * Defines values for XMsAccessSoftDeletedResources. \
+ * {@link KnownXMsAccessSoftDeletedResources} can be used interchangeably with XMsAccessSoftDeletedResources,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **true** \
+ * **false**
+ */
+export type XMsAccessSoftDeletedResources = string;
+
 /** Known values of {@link IdentityType} that the service accepts. */
 export enum KnownIdentityType {
   /** None */
@@ -922,7 +966,7 @@ export enum KnownIdentityType {
   /** SystemAssigned */
   SystemAssigned = "SystemAssigned",
   /** UserAssigned */
-  UserAssigned = "UserAssigned"
+  UserAssigned = "UserAssigned",
 }
 
 /**
@@ -941,7 +985,7 @@ export enum KnownStorageTargetType {
   /** Iscsi */
   Iscsi = "Iscsi",
   /** None */
-  None = "None"
+  None = "None",
 }
 
 /**
@@ -959,7 +1003,7 @@ export enum KnownEncryptionType {
   /** Volume is encrypted at rest with Platform managed key. It is the default encryption type. */
   EncryptionAtRestWithPlatformKey = "EncryptionAtRestWithPlatformKey",
   /** Volume is encrypted at rest with Customer managed key that can be changed and revoked by a customer. */
-  EncryptionAtRestWithCustomerManagedKey = "EncryptionAtRestWithCustomerManagedKey"
+  EncryptionAtRestWithCustomerManagedKey = "EncryptionAtRestWithCustomerManagedKey",
 }
 
 /**
@@ -975,7 +1019,7 @@ export type EncryptionType = string;
 /** Known values of {@link Action} that the service accepts. */
 export enum KnownAction {
   /** Allow */
-  Allow = "Allow"
+  Allow = "Allow",
 }
 
 /**
@@ -986,6 +1030,24 @@ export enum KnownAction {
  * **Allow**
  */
 export type Action = string;
+
+/** Known values of {@link Purge} that the service accepts. */
+export enum KnownPurge {
+  /** True */
+  True = "true",
+  /** False */
+  False = "false",
+}
+
+/**
+ * Defines values for Purge. \
+ * {@link KnownPurge} can be used interchangeably with Purge,
+ *  this enum contains the known values that the service supports.
+ * ### Known values supported by the service
+ * **true** \
+ * **false**
+ */
+export type Purge = string;
 
 /** Known values of {@link VolumeCreateOption} that the service accepts. */
 export enum KnownVolumeCreateOption {
@@ -998,7 +1060,7 @@ export enum KnownVolumeCreateOption {
   /** Disk */
   Disk = "Disk",
   /** DiskRestorePoint */
-  DiskRestorePoint = "DiskRestorePoint"
+  DiskRestorePoint = "DiskRestorePoint",
 }
 
 /**
@@ -1031,7 +1093,7 @@ export enum KnownOperationalStatus {
   /** Stopped */
   Stopped = "Stopped",
   /** StoppedDeallocated */
-  StoppedDeallocated = "Stopped (deallocated)"
+  StoppedDeallocated = "Stopped (deallocated)",
 }
 
 /**
@@ -1055,7 +1117,7 @@ export enum KnownXMsDeleteSnapshots {
   /** True */
   True = "true",
   /** False */
-  False = "false"
+  False = "false",
 }
 
 /**
@@ -1073,7 +1135,7 @@ export enum KnownXMsForceDelete {
   /** True */
   True = "true",
   /** False */
-  False = "false"
+  False = "false",
 }
 
 /**
@@ -1172,7 +1234,10 @@ export type ElasticSansListByResourceGroupNextResponse = ElasticSanList;
 
 /** Optional parameters. */
 export interface VolumeGroupsListByElasticSanOptionalParams
-  extends coreClient.OperationOptions {}
+  extends coreClient.OperationOptions {
+  /** Optional, used to get soft deleted volumes if set to true. */
+  xMsAccessSoftDeletedResources?: XMsAccessSoftDeletedResources;
+}
 
 /** Contains response data for the listByElasticSan operation. */
 export type VolumeGroupsListByElasticSanResponse = VolumeGroupList;
@@ -1204,6 +1269,8 @@ export type VolumeGroupsUpdateResponse = VolumeGroup;
 /** Optional parameters. */
 export interface VolumeGroupsDeleteOptionalParams
   extends coreClient.OperationOptions {
+  /** Optional, used to purge soft deleted volume groups if set to true. */
+  purge?: Purge;
   /** Delay to wait until next poll, in milliseconds. */
   updateIntervalInMs?: number;
   /** A serialized poller which can be used to resume an existing paused Long-Running-Operation. */
@@ -1212,14 +1279,20 @@ export interface VolumeGroupsDeleteOptionalParams
 
 /** Optional parameters. */
 export interface VolumeGroupsGetOptionalParams
-  extends coreClient.OperationOptions {}
+  extends coreClient.OperationOptions {
+  /** Optional, used to get soft deleted volume groups if set to true. */
+  xMsAccessSoftDeletedResources?: XMsAccessSoftDeletedResources;
+}
 
 /** Contains response data for the get operation. */
 export type VolumeGroupsGetResponse = VolumeGroup;
 
 /** Optional parameters. */
 export interface VolumeGroupsListByElasticSanNextOptionalParams
-  extends coreClient.OperationOptions {}
+  extends coreClient.OperationOptions {
+  /** Optional, used to get soft deleted volumes if set to true. */
+  xMsAccessSoftDeletedResources?: XMsAccessSoftDeletedResources;
+}
 
 /** Contains response data for the listByElasticSanNext operation. */
 export type VolumeGroupsListByElasticSanNextResponse = VolumeGroupList;
@@ -1251,6 +1324,8 @@ export type VolumesUpdateResponse = Volume;
 /** Optional parameters. */
 export interface VolumesDeleteOptionalParams
   extends coreClient.OperationOptions {
+  /** Optional, used to purge soft deleted volumes if set to true. */
+  purge?: Purge;
   /** Optional, used to delete snapshots under volume. Allowed value are only true or false. Default value is false. */
   xMsDeleteSnapshots?: XMsDeleteSnapshots;
   /** Optional, used to delete volume if active sessions present. Allowed value are only true or false. Default value is false. */
@@ -1262,24 +1337,45 @@ export interface VolumesDeleteOptionalParams
 }
 
 /** Optional parameters. */
-export interface VolumesGetOptionalParams extends coreClient.OperationOptions {}
+export interface VolumesGetOptionalParams extends coreClient.OperationOptions {
+  /** Optional, used to get soft deleted volumes if set to true. */
+  xMsAccessSoftDeletedResources?: XMsAccessSoftDeletedResources;
+}
 
 /** Contains response data for the get operation. */
 export type VolumesGetResponse = Volume;
 
 /** Optional parameters. */
 export interface VolumesListByVolumeGroupOptionalParams
-  extends coreClient.OperationOptions {}
+  extends coreClient.OperationOptions {
+  /** Optional, used to get soft deleted volumes if set to true. */
+  xMsAccessSoftDeletedResources?: XMsAccessSoftDeletedResources;
+}
 
 /** Contains response data for the listByVolumeGroup operation. */
 export type VolumesListByVolumeGroupResponse = VolumeList;
 
 /** Optional parameters. */
 export interface VolumesListByVolumeGroupNextOptionalParams
-  extends coreClient.OperationOptions {}
+  extends coreClient.OperationOptions {
+  /** Optional, used to get soft deleted volumes if set to true. */
+  xMsAccessSoftDeletedResources?: XMsAccessSoftDeletedResources;
+}
 
 /** Contains response data for the listByVolumeGroupNext operation. */
 export type VolumesListByVolumeGroupNextResponse = VolumeList;
+
+/** Optional parameters. */
+export interface RestoreVolumeOptionalParams
+  extends coreClient.OperationOptions {
+  /** Delay to wait until next poll, in milliseconds. */
+  updateIntervalInMs?: number;
+  /** A serialized poller which can be used to resume an existing paused Long-Running-Operation. */
+  resumeFrom?: string;
+}
+
+/** Contains response data for the restoreVolume operation. */
+export type RestoreVolumeResponse = Volume;
 
 /** Optional parameters. */
 export interface PrivateEndpointConnectionsCreateOptionalParams
@@ -1291,7 +1387,8 @@ export interface PrivateEndpointConnectionsCreateOptionalParams
 }
 
 /** Contains response data for the create operation. */
-export type PrivateEndpointConnectionsCreateResponse = PrivateEndpointConnection;
+export type PrivateEndpointConnectionsCreateResponse =
+  PrivateEndpointConnection;
 
 /** Optional parameters. */
 export interface PrivateEndpointConnectionsGetOptionalParams
@@ -1314,14 +1411,16 @@ export interface PrivateEndpointConnectionsListOptionalParams
   extends coreClient.OperationOptions {}
 
 /** Contains response data for the list operation. */
-export type PrivateEndpointConnectionsListResponse = PrivateEndpointConnectionListResult;
+export type PrivateEndpointConnectionsListResponse =
+  PrivateEndpointConnectionListResult;
 
 /** Optional parameters. */
 export interface PrivateLinkResourcesListByElasticSanOptionalParams
   extends coreClient.OperationOptions {}
 
 /** Contains response data for the listByElasticSan operation. */
-export type PrivateLinkResourcesListByElasticSanResponse = PrivateLinkResourceListResult;
+export type PrivateLinkResourcesListByElasticSanResponse =
+  PrivateLinkResourceListResult;
 
 /** Optional parameters. */
 export interface VolumeSnapshotsListByVolumeGroupOptionalParams